Refactor libs package
All shared codes moves to pkg package
