+++ pkg/entity/auth.go
