+++ pkg/entity/book.go
