+++ pkg/perrors/auth_errors.go
