+++ services/auth/server/router.go
@@ -1,8 +1,8 @@
 package server
 
 import (
-	"github.com/billykore/kore/libs/config"
-	"github.com/billykore/kore/libs/pkg/log"
+	"github.com/billykore/kore/pkg/config"
+	"github.com/billykore/kore/pkg/log"
 	"github.com/billykore/kore/services/auth/service"
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
