+++ services/auth/service/auth_service.go
@@ -1,7 +1,7 @@
 package service
 
 import (
-	"github.com/billykore/kore/libs/entity"
+	"github.com/billykore/kore/pkg/entity"
 	"github.com/billykore/kore/services/auth/usecase"
 	"github.com/labstack/echo/v4"
 )
