+++ services/auth/usecase/auth_usecase.go
@@ -5,16 +5,16 @@ import (
 	"errors"
 	"time"
 
-	"github.com/billykore/kore/libs/entity"
-	"github.com/billykore/kore/libs/model"
-	"github.com/billykore/kore/libs/pkg/codes"
-	"github.com/billykore/kore/libs/pkg/log"
-	"github.com/billykore/kore/libs/pkg/messages"
-	"github.com/billykore/kore/libs/pkg/password"
-	"github.com/billykore/kore/libs/pkg/perrors"
-	"github.com/billykore/kore/libs/pkg/status"
-	"github.com/billykore/kore/libs/pkg/token"
-	"github.com/billykore/kore/libs/repo"
+	"github.com/billykore/kore/pkg/codes"
+	"github.com/billykore/kore/pkg/entity"
+	"github.com/billykore/kore/pkg/log"
+	"github.com/billykore/kore/pkg/messages"
+	"github.com/billykore/kore/pkg/model"
+	"github.com/billykore/kore/pkg/password"
+	"github.com/billykore/kore/pkg/perrors"
+	"github.com/billykore/kore/pkg/repo"
+	"github.com/billykore/kore/pkg/status"
+	"github.com/billykore/kore/pkg/token"
 	"github.com/google/uuid"
 )
