+++ services/todo/cmd/wire_gen.go
@@ -7,9 +7,9 @@
 package main
 
 import (
-	"github.com/billykore/kore/libs/config"
-	"github.com/billykore/kore/libs/db"
-	"github.com/billykore/kore/libs/pkg/log"
+	"github.com/billykore/kore/pkg/config"
+	"github.com/billykore/kore/pkg/db"
+	"github.com/billykore/kore/pkg/log"
 	"github.com/billykore/kore/services/todo/repo"
 	"github.com/billykore/kore/services/todo/server"
 	"github.com/billykore/kore/services/todo/service"
