+++ services/todo/repo/todo_repo.go
@@ -4,9 +4,9 @@ import (
 	"context"
 
 	"cloud.google.com/go/firestore"
-	firestoredb "github.com/billykore/kore/libs/db"
-	"github.com/billykore/kore/libs/model"
-	"github.com/billykore/kore/libs/repo"
+	firestoredb "github.com/billykore/kore/pkg/db"
+	"github.com/billykore/kore/pkg/model"
+	"github.com/billykore/kore/pkg/repo"
 )
 
 type todoRepo struct {
